requirements.txt
Pillow>=9.1.0
tqdm
colorama
humanize